MASSIVE REORG, move all network, keychain, and endpoint data from connect and other utils to separate forlder for extraction later.
