--- libs/utils/src/connect/index.ts
@@ -1,1 +1,0 @@
-export * from './rpcs';
