--- libs/utils/src/connect/rpcs.ts
@@ -1,9 +1,0 @@
-export const HAUS_RPC = {
-  '0x1': `https://${process.env['NX_RIVET_KEY']}.eth.rpc.rivet.cloud/`,
-  '0x5': `https://${process.env['NX_RIVET_KEY']}.goerli.rpc.rivet.cloud/`,
-  '0x64': 'https://rpc.gnosischain.com/',
-  '0xa': 'https://mainnet.optimism.io',
-  '0x89': 'https://polygon-rpc.com/',
-  '0xa4b1': 'https://arb1.arbitrum.io/rpc',
-  '0xa4ec': 'https://forno.celo.org',
-};
