+++ libs/utils/src/index.ts
@@ -2,4 +2,4 @@ export * from './types';
 export * from './constants';
 export * from './content';
 export * from './utils';
-export * from './connect';
+export * from './keychains';
