+++ libs/utils/src/keychains/validNetworks.ts
@@ -1,0 +1,9 @@
+export const VALID_NETWORKS = {
+  '0x1': true,
+  '0x5': true,
+  '0x64': true,
+  '0xa': true,
+  '0x89': true,
+  '0xa4b1': true,
+  '0xa4ec': true,
+};
