+++ libs/utils/src/types/index.ts
@@ -1,4 +1,3 @@
-export * from './keychains';
 export * from './react';
 export * from './contract';
 export * from './general';
