--- libs/utils/src/utils/explorer.ts
@@ -1,12 +1,0 @@
-import { ENDPOINTS } from '../constants';
-import { ValidNetwork } from '../types';
-
-export const generateExplorerLink = ({
-  chainId,
-  address,
-  type = 'address',
-}: {
-  chainId: ValidNetwork;
-  address?: string;
-  type?: string;
-}) => `${ENDPOINTS['EXPLORER'][chainId]}/${type}/${address || ''}`;
